requirements.txt
opencv-python
keyboard
discord.py
